requirements.txt
asgiref==3.5.2
asyncpg==0.27.0
click==8.1.3
colorama==0.4.6
Django==4.1.3
django-async-orm==0.1.10
django-ninja==0.19.1
h11==0.14.0
orjson==3.8.3
psycopg2==2.9.5
pydantic==1.10.2
python-decouple==3.6
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.6
uvicorn==0.20.0
